+++ app/src/main/java/com.jaychang.sac/demo/GithubApi.kt
@@ -4,7 +4,7 @@ import android.net.Uri
 import com.google.gson.annotations.SerializedName
 import com.jaychang.sac.*
 import com.jaychang.sac.annotations.Image
-import com.jaychang.sac.annotations.MockData
+import com.jaychang.sac.annotations.MockResponse
 import com.jaychang.sac.annotations.Unwrap
 import com.jaychang.sac.demo.model.Repo
 import com.jaychang.sac.demo.model.User
@@ -13,7 +13,7 @@ import okhttp3.ResponseBody
 import retrofit2.http.*
 import java.util.concurrent.TimeUnit
 
-class ApiResult<T: Any>: SimpleApiResult<T> {
+class ApiResult<T : Any> : SimpleApiResult<T> {
   @SerializedName("items")
   override lateinit var result: T
 }
@@ -26,21 +26,23 @@ class ApiError : SimpleApiError {
 interface GithubApi {
 
   companion object {
-    fun create() : GithubApi =
-      SimpleApiClient.create<GithubApi, ApiError> {
+    fun create(): GithubApi =
+      SimpleApiClient.create {
         baseUrl = "https://api.github.com"
+        errorClass = ApiError::class
         defaultParameters = mapOf()
         defaultHeaders = mapOf()
         connectTimeout = TimeUnit.MINUTES.toMillis(1)
         readTimeout = TimeUnit.MINUTES.toMillis(1)
         writeTimeout = TimeUnit.MINUTES.toMillis(1)
-        isStethoEnabled = true // default true
-        logLevel = LogLevel.BASIC // default NONE
-        isMockDataEnabled = true // default false
+        isStethoEnabled = true // default: true
+        logLevel = LogLevel.BASIC // default: NONE
+        isMockDataEnabled = true // default: false
         certificatePins = listOf(
           CertificatePin(hostname = "api.foo.com", sha1PublicKeyHash = "0beec7b5ea3f0fdbc95d0dd47f3c5bc275da8a33"),
           CertificatePin(hostname = "api.bar.com", sha256PublicKeyHash = "fcde2b2edba56bf408601fb721fe9b5c338d10ee429ea04fae5511b68fbf8fb9")
         )
+        jsonParser = GsonParser() // default: GsonParser
         errorHandler = { error ->
           when (error) {
             is AuthenticationError -> {}
@@ -55,7 +57,7 @@ interface GithubApi {
 
   @GET("/search/users")
   @Unwrap(ApiResult::class)
-  @MockData(R.raw.get_users)
+  @MockResponse(R.raw.get_users)
   fun getUsers(@Query("q") query: String): Observable<List<User>>
 
   @GET("/repos/{user}/{repo}")