+++ library/src/main/java/com/jaychang/sac/Cancelable.kt
@@ -2,10 +2,10 @@ package com.jaychang.sac
 
 import io.reactivex.disposables.Disposable
 
-class Cancellable(private val disposable: Disposable) {
+class Cancelable(private val disposable: Disposable) {
 
   fun cancel() = disposable.dispose()
 
-  fun isCancelled() = disposable.isDisposed()
+  fun isCanceled() = disposable.isDisposed()
 
 }