+++ library/src/main/java/com/jaychang/sac/ErrorConsumer.kt
@@ -13,11 +13,15 @@ internal class ErrorConsumer<T : Throwable>(private val handler: (Throwable) -> 
       is HttpException -> {
         val code = error.response().code()
         val errorJson = error.response().errorBody()?.string()
-        val message = (ApiManager.gson.fromJson(errorJson!!, ApiManager.apiErrorClass) as SimpleApiError).message
+        val message = if (!errorJson.isNullOrEmpty() && ApiManager.errorClass != null) {
+          (ApiManager.jsonParser.parse<Any>(errorJson!!, ApiManager.errorClass!!.java) as SimpleApiError).message
+        } else {
+          ""
+        }
         when (code) {
           401, 403 -> result = AuthenticationError(code = code, message = message)
           in 400..499 -> result = ClientError(code = code, message = message)
-          in 500..599 -> result = ClientError(code = code, message = message)
+          in 500..599 -> result = ServerError(code = code, message = message)
         }
       }
       is UnknownHostException -> {