+++ library/src/main/java/com/jaychang/sac/annotations/MockData.kt
@@ -1,5 +1,0 @@
-package com.jaychang.sac.annotations
-
-import android.support.annotation.RawRes
-
-annotation class MockData(@RawRes val file: Int)